+++ sdk/communication/communication-chat/test/internal/utils/mockClient.ts
@@ -24,6 +24,14 @@ export const mockParticipant: RestModel.ChatParticipant = {
   shareHistoryTime: new Date("2020-05-26T18:06:06Z"),
 };
 
+export const mockParticipantWithMetadata: RestModel.ChatParticipant = {
+  communicationIdentifier: mockCommunicationIdentifier,
+  displayName: "displayName",
+  metadata: {
+    userType: "C2",
+  },
+};
+
 export const mockSdkModelParticipant: ChatParticipant = {
   id: {
     communicationUserId: mockParticipant.communicationIdentifier.communicationUser?.id as string,
@@ -35,6 +43,7 @@ export const mockSdkModelParticipant: ChatParticipant = {
 export const mockThread: RestModel.ChatThreadProperties = {
   id: "threadid",
   topic: "topic",
+  metadata: { threadType: "primary", secondaryThread: "test-id" },
   createdByCommunicationIdentifier: mockCommunicationIdentifier,
   createdOn: new Date("2020-06-26T18:06:06Z"),
 };
@@ -58,6 +67,81 @@ export const mockMessage: RestModel.ChatMessage = {
   content: {
     message: "content",
     topic: "topic",
+  },
+  createdOn: new Date("2020-06-26T18:06:06Z"),
+  senderDisplayName: "senderDisplayName",
+  senderCommunicationIdentifier: mockCommunicationIdentifier,
+  deletedOn: new Date("2020-06-26T18:06:06Z"),
+  metadata: { tags: "tag" },
+};
+
+export const mockMessageContent: RestModel.ChatMessageContent = {
+  message: "content",
+  topic: "topic",
+  participants: [mockParticipant],
+  attachments: [
+    {
+      id: "id",
+      attachmentType: "teamsInlineImage",
+      contentType: "png",
+      name: "images",
+      url: "url",
+      previewUrl: "previewUrl",
+    },
+    {
+      id: "id1",
+      attachmentType: "teamsImage",
+      contentType: "png",
+      name: "picture1.png",
+      url: "url1",
+      previewUrl: "previewUrl1",
+    },
+    {
+      id: "id2",
+      attachmentType: "file",
+      contentType: "docx",
+      name: "report.docx",
+      url: "url2",
+      previewUrl: "previewUrl2",
+    },
+  ],
+  initiatorCommunicationIdentifier: mockCommunicationIdentifier,
+};
+
+export const mockMessageWithAttachment: RestModel.ChatMessage = {
+  id: "id",
+  type: "text",
+  version: "version",
+  sequenceId: "sequenceId",
+  content: {
+    message: "content",
+    topic: "topic",
+    attachments: [
+      {
+        id: "id",
+        attachmentType: "teamsInlineImage",
+        contentType: "png",
+        name: "images",
+        url: "url",
+        previewUrl: "previewUrl",
+      },
+      {
+        id: "id1",
+        attachmentType: "teamsImage",
+        contentType: "png",
+        name: "picture1.png",
+        url: "url1",
+        previewUrl: "previewUrl1",
+      },
+      {
+        id: "id2",
+        attachmentType: "file",
+        contentType: "docx",
+        name: "report.docx",
+        url: "url2",
+        previewUrl: "previewUrl2",
+      },
+    ],
   },
   createdOn: new Date("2020-06-26T18:06:06Z"),
   senderDisplayName: "senderDisplayName",